ESC-823 introduce dedicated journeys package and move Journey classes there
