+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/journeys/BecomeLeadJourney.scala
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.eusubsidycompliancefrontend.services
+package uk.gov.hmrc.eusubsidycompliancefrontend.journeys
 
 import cats.implicits.catsSyntaxOptionId
 import play.api.libs.json.{Format, Json, OFormat}
 import uk.gov.hmrc.eusubsidycompliancefrontend.controllers.routes
-import uk.gov.hmrc.eusubsidycompliancefrontend.services.BecomeLeadJourney.FormPages._
-import uk.gov.hmrc.eusubsidycompliancefrontend.services.Journey.Form
+import uk.gov.hmrc.eusubsidycompliancefrontend.journeys.BecomeLeadJourney.FormPages._
+import uk.gov.hmrc.eusubsidycompliancefrontend.journeys.Journey.Form
 
 case class BecomeLeadJourney(
   becomeLeadEori: BecomeLeadEoriFormPage = BecomeLeadEoriFormPage(),
