+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/views/BusinessEntityEoriPage.scala.html
@@ -15,7 +15,7 @@
  *@
 
 @import uk.gov.hmrc.eusubsidycompliancefrontend.config.AppConfig
-@import uk.gov.hmrc.eusubsidycompliancefrontend.services.Journey
+@import uk.gov.hmrc.eusubsidycompliancefrontend.journeys.Journey
 @import uk.gov.hmrc.eusubsidycompliancefrontend.controllers
 
 @this(
