+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/views/ConfirmClaimAmountConversionToEuros.scala.html
@@ -16,7 +16,7 @@
 
 @import uk.gov.hmrc.eusubsidycompliancefrontend.controllers
 @import uk.gov.hmrc.eusubsidycompliancefrontend.config.AppConfig
-@import uk.gov.hmrc.eusubsidycompliancefrontend.services.Journey
+@import uk.gov.hmrc.eusubsidycompliancefrontend.journeys.Journey
 @import uk.gov.hmrc.govukfrontend.views.html.components.Text
 @import uk.gov.hmrc.eusubsidycompliancefrontend.views.html.helpers.{H1, P}
