+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/views/InputEmailPage.scala.html
@@ -17,7 +17,7 @@
 @import uk.gov.hmrc.eusubsidycompliancefrontend.controllers
 @import uk.gov.hmrc.eusubsidycompliancefrontend.models.EmailAddress
 @import uk.gov.hmrc.eusubsidycompliancefrontend.config.AppConfig
-@import uk.gov.hmrc.eusubsidycompliancefrontend.services.Journey
+@import uk.gov.hmrc.eusubsidycompliancefrontend.journeys.Journey
 @import uk.gov.hmrc.govukfrontend.views.html.components._
 @import uk.gov.hmrc.eusubsidycompliancefrontend.views.html.helpers.H1
 @import uk.gov.hmrc.eusubsidycompliancefrontend.views.html.helpers.P
