+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/views/NoBusinessPresentPage.scala.html
@@ -16,7 +16,7 @@
 
 @import uk.gov.hmrc.eusubsidycompliancefrontend.controllers
 @import uk.gov.hmrc.eusubsidycompliancefrontend.config.AppConfig
-@import uk.gov.hmrc.eusubsidycompliancefrontend.services.Journey
+@import uk.gov.hmrc.eusubsidycompliancefrontend.journeys.Journey
 
 @this(layout: Layout, formHelper: FormWithCSRF, govukButton: GovukButton, formWithCSRF: FormWithCSRF)
