+++ app/uk/gov/hmrc/eusubsidycompliancefrontend/views/ReportedPaymentsPage.scala.html
@@ -16,7 +16,7 @@
 
 @import uk.gov.hmrc.eusubsidycompliancefrontend.controllers
 @import uk.gov.hmrc.eusubsidycompliancefrontend.config.AppConfig
-@import uk.gov.hmrc.eusubsidycompliancefrontend.services.Journey
+@import uk.gov.hmrc.eusubsidycompliancefrontend.journeys.Journey
 @import uk.gov.hmrc.eusubsidycompliancefrontend.models.Subsidy
 @import uk.gov.hmrc.eusubsidycompliancefrontend.models.UndertakingSubsidies
 @import uk.gov.hmrc.eusubsidycompliancefrontend.models.Undertaking
